feat: move pool header / chart to separate components / add price formatting in tooltip
