+++ apps/marginfi-v2-trading/src/components/common/Pool/index.tsx
@@ -2,3 +2,5 @@ export * from "./BankCard";
 export * from "./CreatePoolDialog";
 export * from "./PoolCard";
 export * from "./PoolSearch";
+export * from "./PoolHeader";
+export * from "./PoolChart";
